[aws-s3] Add support of sdk v3
